CV/pyproject.toml
[project]
name = "cvserve"
version = "0.2.0"
description = "Build styled CV documents from a single YAML source"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "watchdog>=4.0.0",
    "docx2pdf>=0.1.8",
    "typer>=0.15.0",
    "rich>=13.0.0",
]

[tool.setuptools]
packages = ["cvserve"]

[project.scripts]
cvserve = "cvserve:app"
